Optimize chat data seeding with eager loading
This commit updates the database seeding process for the Chat module. In the previous implementation, a lazy loading approach was used which unnecessarily queried the database for each chat message initiated by the user. By switching to eager loading, with the use of 'with' keyword, we reduce the number of queries to the database, improving the efficiency of the seed operation. This results in quicker seeding times and less strain on the database.